services/api/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.5
pydantic==2.8.2
pydantic-settings==2.4.0
sqlalchemy==2.0.32
psycopg[binary]==3.2.1
httpx==0.27.2
boto3==1.35.5
python-multipart==0.0.9
redis==5.0.8
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
